Split the preferences context provider into atomic submodules
