+++ src/global/context/preferences/context.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+
+import type { PreferredLanguageContext } from "./types";
+import { initial } from "./const";
+
+export const Context = createContext<PreferredLanguageContext>(initial);
